Add AVAX C Chain Abandon Seed (#1713)
* Add abandon adress for avax
* Improved typing abandonseed list
