+++ libs/ledger-live-common/src/families/hedera/tests/transaction.test.ts
@@ -25,7 +25,7 @@ const account: Account = {
   blockHeight: 0,
   currency: {
     type: "CryptoCurrency",
-    id: "",
+    id: "hedera",
     managerAppName: "",
     coinType: 0,
     scheme: "",
