+++ libs/ledger-live-common/src/families/hedera/tests/utils.test.ts
@@ -21,7 +21,7 @@ const account: Account = {
   blockHeight: 0,
   currency: {
     type: "CryptoCurrency",
-    id: "",
+    id: "hedera",
     managerAppName: "",
     coinType: 0,
     scheme: "",
