+++ libs/ledgerjs/packages/cryptoassets/src/types.ts
@@ -11,127 +11,3 @@ export type ERC20Token = [
   string,
   string
 ];
-export type CryptoCurrencyId =
-  | "aeternity"
-  | "aion"
-  | "akroma"
-  | "algorand"
-  | "ark"
-  | "atheios"
-  | "avalanche_c_chain"
-  | "banano"
-  | "bitcoin"
-  | "bitcoin_cash"
-  | "bitcoin_gold"
-  | "bitcoin_private"
-  | "bsc"
-  | "callisto"
-  | "cardano"
-  | "cardano_testnet"
-  | "celo"
-  | "clubcoin"
-  | "cosmos"
-  | "cosmos_testnet"
-  | "dash"
-  | "decred"
-  | "dexon"
-  | "ellaism"
-  | "dogecoin"
-  | "digibyte"
-  | "energywebchain"
-  | "eos"
-  | "elastos"
-  | "elrond"
-  | "ethereum"
-  | "ethereum_classic"
-  | "ether1"
-  | "ethergem"
-  | "ethersocial"
-  | "expanse"
-  | "factom"
-  | "fic"
-  | "flow"
-  | "game_credits"
-  | "gochain"
-  | "groestlcoin"
-  | "hcash"
-  | "hedera"
-  | "hedera"
-  | "helium"
-  | "hpb"
-  | "hycon"
-  | "icon"
-  | "icp"
-  | "iota"
-  | "iov"
-  | "kin"
-  | "komodo"
-  | "kusama"
-  | "lbry"
-  | "litecoin"
-  | "lisk"
-  | "mix"
-  | "monero"
-  | "moonriver"
-  | "musicoin"
-  | "nano"
-  | "nem"
-  | "neo"
-  | "nervos"
-  | "nimiq"
-  | "nix"
-  | "nos"
-  | "ontology"
-  | "particl"
-  | "peercoin"
-  | "pirl"
-  | "pivx"
-  | "poa"
-  | "polkadot"
-  | "polygon"
-  | "poswallet"
-  | "qrl"
-  | "qtum"
-  | "ravencoin"
-  | "ripple"
-  | "rise"
-  | "reosc"
-  | "resistance"
-  | "solana"
-  | "stakenet"
-  | "stratis"
-  | "stealthcoin"
-  | "stellar"
-  | "osmosis"
-  | "shyft"
-  | "tezos"
-  | "thundercore"
-  | "tomo"
-  | "tron"
-  | "ubiq"
-  | "vechain"
-  | "vertcoin"
-  | "viacoin"
-  | "wanchain"
-  | "waves"
-  | "zcash"
-  | "zclassic"
-  | "zcoin"
-  | "zencash"
-  | "zilliqa"
-  | "crypto_org"
-  | "bitcoin_testnet"
-  | "ethereum_ropsten"
-  | "ethereum_goerli"
-  | "stacks"
-  | "crypto_org_croeseid"
-  | "solana_testnet"
-  | "solana_devnet"
-  | "filecoin"
-  | "arbitrum"
-  | "cronos"
-  | "fantom"
-  | "flare"
-  | "songbird"
-  | "moonbeam"
-  | "near";
