+++ libs/ledgerjs/packages/types-cryptoassets/src/index.ts
@@ -1,5 +1,131 @@
 import { CoinType } from "./slip44";
 
+// All cryptocurrency ids
+// list should be append only. **do not modify existing ids**
+export type CryptoCurrencyId =
+  | "aeternity"
+  | "aion"
+  | "akroma"
+  | "algorand"
+  | "ark"
+  | "atheios"
+  | "avalanche_c_chain"
+  | "banano"
+  | "bitcoin"
+  | "bitcoin_cash"
+  | "bitcoin_gold"
+  | "bitcoin_private"
+  | "bsc"
+  | "callisto"
+  | "cardano"
+  | "cardano_testnet"
+  | "celo"
+  | "clubcoin"
+  | "cosmos"
+  | "cosmos_testnet"
+  | "dash"
+  | "decred"
+  | "dexon"
+  | "ellaism"
+  | "dogecoin"
+  | "digibyte"
+  | "energywebchain"
+  | "eos"
+  | "elastos"
+  | "elrond"
+  | "ethereum"
+  | "ethereum_classic"
+  | "ether1"
+  | "ethergem"
+  | "ethersocial"
+  | "expanse"
+  | "factom"
+  | "fic"
+  | "flow"
+  | "game_credits"
+  | "gochain"
+  | "groestlcoin"
+  | "hcash"
+  | "hedera"
+  | "helium"
+  | "hpb"
+  | "hycon"
+  | "icon"
+  | "icp"
+  | "iota"
+  | "iov"
+  | "kin"
+  | "komodo"
+  | "kusama"
+  | "lbry"
+  | "litecoin"
+  | "lisk"
+  | "mix"
+  | "monero"
+  | "moonriver"
+  | "musicoin"
+  | "nano"
+  | "nem"
+  | "neo"
+  | "nervos"
+  | "nimiq"
+  | "nix"
+  | "nos"
+  | "ontology"
+  | "particl"
+  | "peercoin"
+  | "pirl"
+  | "pivx"
+  | "poa"
+  | "polkadot"
+  | "polygon"
+  | "poswallet"
+  | "qrl"
+  | "qtum"
+  | "ravencoin"
+  | "ripple"
+  | "rise"
+  | "reosc"
+  | "resistance"
+  | "solana"
+  | "stakenet"
+  | "stratis"
+  | "stealthcoin"
+  | "stellar"
+  | "osmosis"
+  | "shyft"
+  | "tezos"
+  | "thundercore"
+  | "tomo"
+  | "tron"
+  | "ubiq"
+  | "vechain"
+  | "vertcoin"
+  | "viacoin"
+  | "wanchain"
+  | "waves"
+  | "zcash"
+  | "zclassic"
+  | "zcoin"
+  | "zencash"
+  | "zilliqa"
+  | "crypto_org"
+  | "bitcoin_testnet"
+  | "ethereum_ropsten"
+  | "ethereum_goerli"
+  | "stacks"
+  | "crypto_org_croeseid"
+  | "solana_testnet"
+  | "solana_devnet"
+  | "filecoin"
+  | "arbitrum"
+  | "cronos"
+  | "fantom"
+  | "flare"
+  | "songbird"
+  | "moonbeam"
+  | "near";
+
 /**
  *
  */
@@ -76,7 +202,9 @@ export type ExplorerView = {
 export type CryptoCurrency = CurrencyCommon & {
   type: "CryptoCurrency";
   // unique internal id of a crypto currency
-  id: string;
+  // LBRY | groestcoin | osmo are expectional ids
+  // because in `cryptocurrenciesById` the key of the object is different from the id
+  id: CryptoCurrencyId | "LBRY" | "groestcoin" | "osmo";
   // define if a crypto is a fork from another coin. helps dealing with split/unsplit
   forkedFrom?: string;
   // name of the app as shown in the Manager
